Handle garbage in hash config.
