+++ lib/reek/errors/bad_detector_configuration_in_comment_error.rb
@@ -1,0 +1,22 @@
+# frozen_string_literal: true
+module Reek
+  module Errors
+    # TODO
+    class BadDetectorConfigurationInCommentError < RuntimeError
+      BAD_DETECTOR_CONFIGURATION_MESSAGE = <<-EOS.freeze
+
+      TODO %s %s %s %s
+
+      EOS
+
+      def initialize(detector:, source:, line:, original_comment:)
+        message = format(BAD_DETECTOR_CONFIGURATION_MESSAGE,
+                         detector,
+                         source,
+                         line,
+                         original_comment)
+        super message
+      end
+    end
+  end
+end
